+++ Code/Framework/AzToolsFramework/Tests/Prefab/Benchmark/PrefabCreateBenchmarks.cpp
@@ -16,7 +16,7 @@ namespace Benchmark
 
     BENCHMARK_DEFINE_F(BM_PrefabCreate, CreatePrefabs_SingleEntityEach)(::benchmark::State& state)
     {
-        const unsigned int numEntities = state.range();
+        const unsigned int numEntities = static_cast<unsigned int>(state.range());
         const unsigned int numInstances = numEntities;
 
         CreateFakePaths(numInstances);
@@ -58,7 +58,7 @@ namespace Benchmark
 
     BENCHMARK_DEFINE_F(BM_PrefabCreate, CreatePrefab_FromEntities)(::benchmark::State& state)
     {
-        const unsigned int numEntities = state.range();
+        const unsigned int numEntities = static_cast<unsigned int>(state.range());
 
         for (auto _ : state)
         {
@@ -72,7 +72,7 @@ namespace Benchmark
             AZStd::unique_ptr<Instance> instance = m_prefabSystemComponent->CreatePrefab(
                 entities
                 , {}
-            , m_pathString);
+                , m_pathString);
 
             state.PauseTiming();
 
@@ -93,7 +93,7 @@ namespace Benchmark
 
     BENCHMARK_DEFINE_F(BM_PrefabCreate, CreatePrefab_FromSingleDepthInstances)(::benchmark::State& state)
     {
-        const unsigned int numInstancesToAdd = state.range();
+        const unsigned int numInstancesToAdd = static_cast<unsigned int>(state.range());
         const unsigned int numEntities = numInstancesToAdd;
 
         // Create fake paths for all the nested instances
@@ -144,7 +144,7 @@ namespace Benchmark
 
     BENCHMARK_DEFINE_F(BM_PrefabCreate, CreatePrefab_FromLinearNestingOfInstances)(::benchmark::State& state)
     {
-        const unsigned int numInstances = state.range();
+        const unsigned int numInstances = static_cast<unsigned int>(state.range());
 
         // Create fake paths for all the nested instances
         // plus the root instance
@@ -165,7 +165,7 @@ namespace Benchmark
             {
                 nestedInstanceRoot = m_prefabSystemComponent->CreatePrefab(
                     {},
-                    MakeInstanceList( AZStd::move(nestedInstanceRoot) ),
+                    MakeInstanceList(AZStd::move(nestedInstanceRoot)),
                     m_paths[instanceCounter]);
             }
