+++ Code/Framework/AzToolsFramework/Tests/Prefab/Benchmark/SpawnableCreateBenchmarks.cpp
@@ -18,7 +18,7 @@ namespace Benchmark
 
     BENCHMARK_DEFINE_F(BM_SpawnableCreate, CreateSpawnable_SingleEntityInstance)(::benchmark::State& state)
     {
-        const unsigned int numSpawnables = state.range();
+        const unsigned int numSpawnables = static_cast<unsigned int>(state.range());
 
         AZStd::unique_ptr<Instance> instance(m_prefabSystemComponent->CreatePrefab(
             { CreateEntity("Entity1") },
