+++ Gems/AudioSystem/Code/Source/Editor/QConnectionsWidget.cpp
@@ -138,7 +138,7 @@ namespace AudioControls
                 for (int i = 0; i < size; ++i)
                 {
                     QListWidgetItem* listItem = m_connectionList->item(i);
-                    if (listItem && listItem->data(eMDR_ID).toInt() == middlewareControl->GetId())
+                    if (listItem && listItem->data(eMDR_ID).toInt() == static_cast<int>(middlewareControl->GetId()))
                     {
                         m_connectionList->clearSelection();
                         listItem->setSelected(true);
